+++ backend/canisters/community/impl/src/updates/update_community.rs
@@ -12,7 +12,9 @@ use types::{
     OptionalCommunityPermissions, PrimaryLanguageChanged, Timestamped, UserId,
 };
 use utils::document_validation::{validate_avatar, validate_banner};
-use utils::group_validation::{validate_description, validate_name, validate_rules, NameValidationError, RulesValidationError};
+use utils::text_validation::{
+    validate_community_name, validate_description, validate_rules, NameValidationError, RulesValidationError,
+};
 
 #[update]
 #[trace]
@@ -115,7 +117,7 @@ fn prepare(args: &Args, state: &RuntimeState) -> Result<PrepareResult, Response>
     let gate = args.gate.as_ref().apply_to(state.data.gate.value.as_ref());
 
     if let Some(name) = &args.name {
-        if let Err(error) = validate_name(name, state.data.is_public) {
+        if let Err(error) = validate_community_name(name, state.data.is_public) {
             return Err(match error {
                 NameValidationError::TooShort(s) => NameTooShort(s),
                 NameValidationError::TooLong(l) => NameTooLong(l),
@@ -344,5 +346,6 @@ fn merge_permissions(new: OptionalCommunityPermissions, old: &CommunityPermissio
         update_details: new.update_details.unwrap_or(old.update_details),
         create_public_channel: new.create_public_channel.unwrap_or(old.create_public_channel),
         create_private_channel: new.create_private_channel.unwrap_or(old.create_private_channel),
+        manage_user_groups: new.manage_user_groups.unwrap_or(old.manage_user_groups),
     }
 }