+++ backend/integration_tests/src/change_group_role_tests.rs
@@ -41,7 +41,6 @@ fn init_test_data(env: &mut StateMachine, canister_ids: &CanisterIds, controller
     TestData { user1, user2, group_id }
 }
 
-#[allow(dead_code)]
 struct TestData {
     user1: User,
     user2: User,