+++ backend/integration_tests/src/platform_moderator_tests.rs
@@ -111,7 +111,7 @@ fn report_message_succeeds() {
 
 fn init_test_data(env: &mut StateMachine, canister_ids: &CanisterIds, controller: Principal) -> TestData {
     let user1 = client::register_diamond_user(env, canister_ids, controller);
-    let user2 = client::local_user_index::happy_path::register_user(env, canister_ids.local_user_index);
+    let user2 = client::register_diamond_user(env, canister_ids, controller);
 
     client::user_index::add_platform_moderator(
         env,
@@ -140,7 +140,6 @@ fn init_test_data(env: &mut StateMachine, canister_ids: &CanisterIds, controller
     }
 }
 
-#[allow(dead_code)]
 struct TestData {
     user1: User,
     user2: User,