+++ backend/libraries/types/src/user_groups.rs
@@ -1,0 +1,16 @@
+use crate::UserId;
+use candid::CandidType;
+use serde::{Deserialize, Serialize};
+
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+pub struct UserGroupSummary {
+    pub user_group_id: u32,
+    pub name: String,
+    pub members: u32,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+pub struct UserGroupMembers {
+    pub user_group_id: u32,
+    pub members: Vec<UserId>,
+}