+++ frontend/app/src/components/App.svelte
@@ -550,6 +550,8 @@
             display: flex;
             height: 100vh;
             height: calc(var(--vh, 1vh) * 100);
+            height: 100dvh; // firefox will ignore this
+            position: fixed;
 
             @include size-below(lg) {
                 padding: $sp3;
@@ -571,6 +573,7 @@
                 color: var(--landing-txt);
                 min-height: 100vh;
                 height: unset;
+                position: unset;
             }
 
             @media (hover: none) {