+++ frontend/app/src/components/home/CurrentChat.svelte
@@ -338,8 +338,6 @@
             editingEvent={$currentChatEditingEvent}
             replyingTo={$currentChatReplyingTo}
             textContent={$currentChatTextContent}
-            members={$currentChatMembers}
-            blockedUsers={$currentChatBlockedUsers}
             {user}
             mode={"message"}
             {joining}