+++ frontend/app/src/components/home/CurrentChatSearchHeader.svelte
@@ -5,7 +5,7 @@
     import ChevronUp from "svelte-material-icons/ChevronUp.svelte";
     import ChevronDown from "svelte-material-icons/ChevronDown.svelte";
     import Close from "svelte-material-icons/Close.svelte";
-    import type { MessageMatch, ChatSummary, OpenChat } from "openchat-client";
+    import type { MessageMatch, ChatSummary, OpenChat, UserSummary } from "openchat-client";
     import HoverIcon from "../HoverIcon.svelte";
     import { iconSize } from "../../stores/iconSize";
     import MentionPicker from "./MentionPicker.svelte";
@@ -15,7 +15,6 @@
 
     const dispatch = createEventDispatcher();
     const client = getContext<OpenChat>("client");
-    const reverseUserLookup: Record<string, string> = {};
     const mentionRegex = /@([\d\w_]*)$/;
 
     let lastSearchTerm = "";
@@ -28,12 +27,10 @@
     let searchBoxHeight: number | undefined;
     let rangeToReplace: [number, number] | undefined = undefined;
     let timer: number | undefined;
+    let userLookupByUsername: Record<string, UserSummary> | undefined = undefined;
 
-    $: userStore = client.userStore;
     $: count = matches.length > 0 ? `${currentMatch + 1}/${matches.length}` : "";
     $: isGroup = chat.kind === "group_chat";
-    $: members = client.currentChatMembers;
-    $: blockedUsers = client.currentChatBlockedUsers;
 
     onMount(() => {
         inputElement.focus();
@@ -116,9 +113,9 @@
 
         let mentionedSet = new Set<string>();
         let expandedText = text.replace(/@([\w\d_]*)/g, (match, p1) => {
-            const userId = reverseUserLookup[p1];
-            if (userId !== undefined) {
-                mentionedSet.add(userId);
+            const user = getOrBuildUserLookupByUsername()[p1.toLowerCase()];
+            if (user !== undefined) {
+                mentionedSet.add(user.userId);
                 return "";
             } else {
                 console.log(
@@ -186,6 +183,7 @@
             if (matches.index !== undefined) {
                 rangeToReplace = [matches.index, pos];
                 mentionPrefix = matches[1].toLowerCase() || undefined;
+                getOrBuildUserLookupByUsername();
                 showMentionPicker = true;
             }
         } else {
@@ -225,17 +223,13 @@
         searchTerm = inputElement.value;
     }
 
-    function mention(ev: CustomEvent<string>): void {
-        const user = $userStore[ev.detail];
-        const username = user?.username ?? $_("unknown");
-        const userLabel = `@${username}`;
+    function mention(ev: CustomEvent<UserSummary>): void {
+        const user = ev.detail;
+        const userLabel = `@${user.username}`;
 
         replaceTextWith(userLabel);
 
         showMentionPicker = false;
-        if (user !== undefined) {
-            reverseUserLookup[username] = user.userId;
-        }
         performSearch();
     }
 
@@ -243,18 +237,23 @@
         showMentionPicker = false;
         setCaretToEnd();
     }
+
+    function getOrBuildUserLookupByUsername(): Record<string, UserSummary> {
+        if (userLookupByUsername === undefined) {
+            userLookupByUsername = client.buildUserLookupForMentions(true);
+        }
+        return userLookupByUsername;
+    }
 </script>
 
 <svelte:window on:keydown={onWindowKeyDown} />
 
 {#if showMentionPicker}
     <MentionPicker
+        {userLookupByUsername}
         offset={searchBoxHeight ?? 80}
         direction={"down"}
-        mentionSelf
         prefix={mentionPrefix}
-        members={$members}
-        blockedUsers={$blockedUsers}
         on:close={cancelMention}
         on:mention={mention} />
 {/if}