+++ frontend/app/src/components/home/MentionPicker.svelte
@@ -3,23 +3,19 @@
     import Menu from "../Menu.svelte";
     import VirtualList from "../VirtualList.svelte";
 
-    import type { Member, PartialUserSummary, OpenChat } from "openchat-client";
+    import type { UserSummary, OpenChat } from "openchat-client";
     import { createEventDispatcher, getContext } from "svelte";
-    import { _ } from "svelte-i18n";
     import Avatar from "../Avatar.svelte";
     import { AvatarSize } from "openchat-client";
     import { mobileWidth } from "../../stores/screenDimensions";
 
     const client = getContext<OpenChat>("client");
-    const user = client.user;
 
-    export let blockedUsers: Set<string>;
-    export let members: Member[];
+    export let userLookupByUsername: Record<string, UserSummary>;
     export let prefix: string | undefined;
     export let offset: number;
     export let direction: "up" | "down" = "up";
     export let border = false;
-    export let mentionSelf = false;
 
     let index = 0;
     $: userStore = client.userStore;
@@ -28,22 +24,9 @@
     $: maxHeight =
         direction === "down" ? `${3.2 * itemHeight + borderWidth}px` : "calc(var(--vh, 1vh) * 50)";
 
-    $: unblocked = members.filter(
-        (m) => !blockedUsers.has(m.userId) && (mentionSelf || m.userId !== user.userId)
-    );
-
-    $: reverseLookup = unblocked.reduce((lookup, u) => {
-        const user = $userStore[u.userId];
-        if (user !== undefined && user.username !== undefined) {
-            lookup[user.username.toLowerCase()] = user;
-        }
-        return lookup;
-    }, {} as Record<string, PartialUserSummary>);
-
-    $: filtered = unblocked.filter(
-        (p) =>
-            prefix === undefined ||
-            $userStore[p.userId]?.username?.toLowerCase().startsWith(prefix?.toLowerCase())
+    $: prefixLower = prefix?.toLowerCase();
+    $: filtered = Object.values(userLookupByUsername).filter(
+        (user) => (prefixLower === undefined || user.username.toLowerCase().startsWith(prefixLower))
     );
 
     $: style =
@@ -57,12 +40,8 @@
 
     const dispatch = createEventDispatcher();
 
-    export function userFromUsername(username: string): PartialUserSummary | undefined {
-        return reverseLookup[username.toLowerCase()];
-    }
-
-    function mention(userId: string) {
-        dispatch("mention", userId);
+    function mention(user: UserSummary) {
+        dispatch("mention", user);
     }
 
     function onKeyDown(ev: KeyboardEvent): void {
@@ -85,7 +64,7 @@
             case "Enter":
                 const user = filtered[index];
                 if (user) {
-                    mention(user.userId);
+                    mention(user);
                 }
                 ev.preventDefault();
                 ev.stopPropagation();
@@ -102,7 +81,7 @@
     {style}>
     <Menu>
         <VirtualList keyFn={(p) => p.userId} items={filtered} let:item let:itemIndex>
-            <MenuItem selected={itemIndex === index} on:click={() => mention(item.userId)}>
+            <MenuItem selected={itemIndex === index} on:click={() => mention(item)}>
                 <div class="avatar" slot="icon">
                     <Avatar
                         url={client.userAvatarUrl($userStore[item.userId])}
@@ -110,7 +89,7 @@
                         size={AvatarSize.Small} />
                 </div>
                 <div slot="text">
-                    {$userStore[item.userId]?.username ?? $_("unknown")}
+                    {item.username}
                 </div>
             </MenuItem>
         </VirtualList>