+++ frontend/openchat-agent/src/services/userIndex/candid/types.d.ts
@@ -248,6 +248,7 @@ export interface CommunityCanisterCommunitySummary {
   'description' : string,
   'last_updated' : TimestampMillis,
   'channels' : Array<CommunityCanisterChannelSummary>,
+  'user_groups' : Array<UserGroup>,
   'avatar_id' : [] | [bigint],
   'membership' : [] | [CommunityMembership],
   'frozen' : [] | [FrozenGroupInfo],
@@ -267,6 +268,7 @@ export interface CommunityCanisterCommunitySummaryUpdates {
   'description' : [] | [string],
   'last_updated' : TimestampMillis,
   'channels_removed' : Array<ChannelId>,
+  'user_groups' : Array<UserGroup>,
   'avatar_id' : DocumentIdUpdate,
   'channels_added' : Array<CommunityCanisterChannelSummary>,
   'membership' : [] | [CommunityMembershipUpdates],
@@ -305,6 +307,7 @@ export type CommunityPermissionRole = { 'Owners' : null } |
   { 'Members' : null };
 export interface CommunityPermissions {
   'create_public_channel' : CommunityPermissionRole,
+  'manage_user_groups' : CommunityPermissionRole,
   'update_details' : CommunityPermissionRole,
   'remove_members' : CommunityPermissionRole,
   'invite_users' : CommunityPermissionRole,
@@ -926,6 +929,7 @@ export interface NotificationEnvelope {
 }
 export interface OptionalCommunityPermissions {
   'create_public_channel' : [] | [CommunityPermissionRole],
+  'manage_user_groups' : [] | [CommunityPermissionRole],
   'update_details' : [] | [CommunityPermissionRole],
   'remove_members' : [] | [CommunityPermissionRole],
   'invite_users' : [] | [CommunityPermissionRole],
@@ -1275,6 +1279,11 @@ export interface User { 'username' : string, 'user_id' : UserId }
 export interface UserArgs {
   'username' : [] | [string],
   'user_id' : [] | [UserId],
+}
+export interface UserGroup {
+  'members' : number,
+  'name' : string,
+  'user_group_id' : number,
 }
 export type UserId = CanisterId;
 export type UserRegistrationCanisterResponse = { 'Success' : CanisterId } |