+++ frontend/openchat-client/src/openchat.ts
@@ -316,6 +316,7 @@ import type {
     CryptocurrencyContent,
     CryptocurrencyDetails,
     CryptocurrencyTransfer,
+    Member,
     Mention,
 } from "openchat-shared";
 import {
@@ -1128,6 +1129,9 @@ export class OpenChat extends OpenChatAgentWorker {
     }
 
     setCommunityIndexes(indexes: Record<string, number>): Promise<boolean> {
+        Object.entries(indexes).forEach(([k, v]) =>
+            localCommunitySummaryUpdates.updateIndex({ kind: "community", communityId: k }, v),
+        );
         return this.sendRequest({ kind: "setCommunityIndexes", indexes }).catch((err) => {
             this._logger.error("Failed to set community indexes: ", err);
             return false;
@@ -4565,6 +4569,22 @@ export class OpenChat extends OpenChatAgentWorker {
                 }),
             ),
         );
+    }
+
+    buildUserLookupForMentions(includeSelf: boolean): Record<string, UserSummary> {
+        const map = {} as Record<string, UserSummary>;
+        const userStore = this._liveState.userStore;
+        const currentUser = this._user?.userId;
+        for (const member of this._liveState.currentChatMembers) {
+            const userId = member.userId;
+            if (includeSelf || userId !== currentUser) {
+                const user = userStore[userId];
+                if (user !== undefined && user.username !== undefined) {
+                    map[user.username.toLowerCase()] = user as UserSummary;
+                }
+            }
+        }
+        return map;
     }
 
     // **** Communities Stuff